Remove old codes
