+++ Classes/Session/SessionMerge.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Hi3Helper.Http
 {
-    public partial class HttpNew
+    public partial class Http
     {
         public async Task Merge()
         {
