+++ Classes/Session/SessionVerification.cs
@@ -6,7 +6,7 @@ using System.Text;
 
 namespace Hi3Helper.Http
 {
-    public partial class HttpNew
+    public partial class Http
     {
         /*
         public async Task TryRunSessionVerification()
